Moved classes to appropriate package.
The ServerFinder class and ServerDiscoveryHandler class have been moved
out of the redundant clientside package.
